+++ Sources/SuperwallKit/StoreKit/Products/Receipt Manager/Receipt Manager/ReceiptManager.swift
@@ -76,6 +76,7 @@ actor ReceiptManager {
         case .verified(let transaction),
           .unverified(let transaction, _):
           Self.appTransactionId = transaction.appTransactionID
+          Superwall.shared.dequeueAttributionProps()
         }
       }
     }